bugfix: fixed glitches with File and Folder selectors
- Folder selector needs a slash in the end of the path, otherwise it goes to the parent path. This happened in the Create Hardfile and Disk Swapper panels.
- File selector in Floppy and Quickstart panels would show incorrect directory contents, if it started with a filename/path that could not be resolved

+++ src/osdep/gui/PanelQuickstart.cpp
@@ -72,7 +72,7 @@ public:
 
 	int getNumberOfElements() override
 	{
-		return types.size();
+		return static_cast<int>(types.size());
 	}
 
 	std::string getElementAt(int i) override
